+++ BranchSDK/test/SleeperTest.cpp
@@ -1,4 +1,4 @@
-#include <BranchIO/Sleeper.h>
+#include <BranchIO/Internal/Sleeper.h>
 #include <gmock/gmock.h>
 #include <Poco/Clock.h>
 #include <Poco/Timer.h>
